queue-service/pyproject.toml
[tool.poetry]
name = "queue-service"
version = "0.1.0"
description = ""
authors = ["Daniel N. Rebouças <danielnreboucas@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
flask = "^3.0.3"
jsonschema = "^4.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
